Fix cli node_access_rebuild error
When you add domains during a drush site install, you need to call node_access_rebuild. This calls domain_access_node_grants, and because the domain_get_domain function is url based you have to create a localhost domain because that is the default. 

To prevent the creation of a domain just because it's the default value, I added an extra check that gets the default domain instead. it doesn't interfere with the permission rebuild via the web interface.