Currently untested!
Added requested documentation;
Added additional validation to domain:add and to domain:scheme;
Modified domain:scheme to return the current scheme or use a --set option to change it.
Added some additional type specifiers, and changed variations on 'domain' to be 'domain';
Probably corrected validate domain code;
Deleted the command to change the machine name, now that is the same as the id.